refactor: rename collection module to collection_reader
