+++ tdp/core/collection_reader.py
